+++ lib/src/flutter/navigation.dart
@@ -11,29 +11,19 @@
  * limitations under the License.
  */
 
-import 'package:test/test.dart';
-import 'package:dart_extensions/src/date.dart';
+import 'package:flutter/material.dart';
 
-main() {
-  group('date time', () {
-    test('toMilliseconds', () {
-      expect(1.toMilliseconds(), Duration(milliseconds: 1));
-    });
+extension NavigationExtensions on State {
+  /// Navigate to another widget
+  Future<T> navigateTo<T>({@required Route<T> route}) => Navigator.push(this.context, route);
 
-    test('toSeconds', () {
-      expect(1.toSeconds(), Duration(seconds: 1));
-    });
+  /// Navigate back to the last widget
+  bool navigateBack<T>({T result}) => Navigator.pop(context, result);
 
-    test('toMinutes', () {
-      expect(1.toMinutes(), Duration(minutes: 1));
-    });
 
-    test('toHours', () {
-      expect(1.toHours(), Duration(hours: 1));
-    });
+  /// Navigate to widget by the route name
+  Future<T> navigateByRouteName<T>(String routeName, {Object args}) =>
+      Navigator.pushNamed(context, routeName, arguments: args);
 
-    test('toDays', () {
-      expect(1.toDays(), Duration(days: 1));
-    });
-  });
 }
+