+++ lib/src/flutter/padding.dart
@@ -1,0 +1,46 @@
+/*
+ * Copyright 2020 Idan Ayalon. All rights reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import 'package:flutter/material.dart';
+
+extension PaddingExtensions on Widget {
+  Padding paddingAll(double value, {Key key}) {
+    return Padding(
+      key: key,
+      padding: EdgeInsets.all(value),
+      child: this,
+    );
+  }
+
+  Padding paddingLTRB(
+    double left,
+    double top,
+    double right,
+    double bottom, {
+    Key key,
+  }) =>
+      Padding(
+        key: key,
+        padding: EdgeInsets.fromLTRB(left, top, right, bottom),
+        child: this,
+      );
+
+  Padding paddingSymmetric({Key key, double v = 0.0, double h = 0.0}) => Padding(
+        key: key,
+        padding: EdgeInsets.symmetric(
+          vertical: v,
+          horizontal: h,
+        ),
+        child: this,
+      );
+}