+++ lib/src/flutter/transforms.dart
@@ -10,6 +10,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-main() {
 
+import 'package:dart_extensions/src/flutter/transforms/click_translate.dart';
+import 'package:flutter/material.dart';
+
+extension TransformExtensions on Widget {
+  Widget get pushEffectOnClick => TranslateOnClick(child: this);
 }