+++ lib/src/flutter/transforms/click_translate.dart
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2020 Idan Ayalon. All rights reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import 'package:flutter/material.dart';
+
+class TranslateOnClick extends StatefulWidget {
+  final Widget child;
+
+  const TranslateOnClick({Key key, this.child}) : super(key: key);
+
+  @override
+  _TranslateOnClickState createState() => _TranslateOnClickState();
+}
+
+class _TranslateOnClickState extends State<TranslateOnClick> {
+  final nonClickTransform = Matrix4.identity();
+  final clickTransform = Matrix4.identity()..translate(0, -10, 0);
+
+  bool _clicking = false;
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTapDown: (d) => _userClick(true),
+      onTapUp: (d) => _userClick(false),
+      child: AnimatedContainer(
+        duration: Duration(milliseconds: 200),
+        child: widget.child,
+        transform: _clicking ? clickTransform : nonClickTransform,
+      ),
+    );
+  }
+
+  void _userClick(bool click) {
+    setState(() {
+      _clicking = click;
+    });
+  }
+}