+++ lib/src/http.dart
@@ -11,9 +11,11 @@
  * limitations under the License.
  */
 
-import 'package:http/http.dart' as http;
 import 'dart:convert' as convert;
-import 'strings.dart';
+
+import 'package:http/http.dart' as http;
+
+import 'package:dart_extensions/src/strings.dart';
 
 const _defaultHeaders = {"Content-type": "application/json"};
 
@@ -21,7 +23,7 @@ extension HttpExtensions on String {
   /// Sends an HTTP GET request with the given headers to the given URL, which can
   /// be a [Uri] or a [String].
   Future<dynamic> httpGet() async {
-    if (this.isNullOrEmpty) return;
+    if (this.isEmptyOrNull) return;
 
     try {
       final response = await http.get(this);
@@ -37,7 +39,7 @@ extension HttpExtensions on String {
   /// which can be a [Uri] or a [String].
   Future<dynamic> httpPost(String json,
       [Map<String, String> headers = _defaultHeaders]) async {
-    if (this.isNullOrEmpty) return;
+    if (this.isEmptyOrNull) return;
 
     try {
       final response = await http.post(this, headers: headers, body: json);
@@ -53,7 +55,7 @@ extension HttpExtensions on String {
   /// which can be a [Uri] or a [String].
   Future<dynamic> httpPut(String json,
       [Map<String, String> headers = _defaultHeaders]) async {
-    if (this.isNullOrEmpty) return;
+    if (this.isEmptyOrNull) return;
 
     try {
       final response = await http.put(this, headers: headers, body: json);
@@ -68,7 +70,7 @@ extension HttpExtensions on String {
   /// Sends an HTTP DELETE request with the given headers to the given URL, which
   /// can be a [Uri] or a [String].
   Future<dynamic> httpDelete({Map<String, String> headers}) async {
-    if (this.isNullOrEmpty) return;
+    if (this.isEmptyOrNull) return;
 
     try {
       final response = await http.delete(this, headers: headers);