+++ test/int_test.dart
@@ -11,10 +11,17 @@
  * limitations under the License.
  */
 import 'package:test/test.dart';
-import 'package:dart_extensions/int.dart';
+import 'package:dart_extensions/src/int.dart';
 
 main() {
   group('integers', () {
+
+    test('inRangeOf', () {
+      expect(1.inRangeOf(0, 3), 1);
+      expect(2.inRangeOf(3, 4), 3);
+      expect(5.inRangeOf(3, 4), 4);
+    });
+
 
     test('isEven', () {
       expect(2.isEven, true);